add DirectorNotFoundException to director

+++ src/main/java/com/codecool/VideoStoreRestApi/services/DirectorService.java
@@ -1,5 +1,6 @@
 package com.codecool.VideoStoreRestApi.services;
 
+import com.codecool.VideoStoreRestApi.Exceptions.DirectorNotFoundException;
 import com.codecool.VideoStoreRestApi.model.Director;
 import com.codecool.VideoStoreRestApi.model.Movie;
 import com.codecool.VideoStoreRestApi.repositories.DirectorRepository;
@@ -21,23 +22,32 @@ public class DirectorService {
     }
 
     public Collection<Director> getAllDirectors() {
-        return this.directorRepository.findAll();
+        Collection<Director> directorsList = this.directorRepository.findAll();
+        if(directorsList.isEmpty()) throw new DirectorNotFoundException("no any director was found");
+        return directorsList;
     }
 
     public Director getDirectorById(Long id) {
-        return this.directorRepository.findOne(id);
+        Director director = this.directorRepository.findOne(id);
+        if(director == null) throw new DirectorNotFoundException("director was found");
+        return director;
     }
 
     public void deleteDirectorById(Long idDirector) {
+        Director director = getDirectorById(idDirector);
+        if(director == null) throw new DirectorNotFoundException("director was found");
         this.directorRepository.delete(idDirector);
     }
 
     public void deleteAllDirectors() {
+        Collection<Director> directorsList = this.directorRepository.findAll();
+        if(directorsList.isEmpty()) throw new DirectorNotFoundException("no any director was found");
         this.directorRepository.deleteAll();
     }
 
     public void updateDirector(Long id, String first_name, String last_name) {
         Director director = getDirectorById(id);
+        if(director == null) throw new DirectorNotFoundException("director was found");
         if(first_name != null){
             director.setFirstName(first_name);
         }
@@ -59,17 +69,20 @@ public class DirectorService {
     this.directorRepository.save(director);
     }
 
-
     public boolean isDirectorExist(String firstName, String lastName){
         return this.directorRepository.existsDirectorByFirstNameAndLastName(firstName, lastName);
     }
 
     public Director getDirector(String firstName, String lastName) {
-        return this.directorRepository.getDirectorByFirstNameAndLastName(firstName, lastName);
+        Director director = this.directorRepository.getDirectorByFirstNameAndLastName(firstName, lastName);
+        if(director == null) throw new DirectorNotFoundException("director was found");
+        return director;
     }
 
     public Collection<Movie> getDirectorMovies(Long idDirector){
-        return this.movieRepository.findMovieByDirectorsIdDirector(idDirector);
+        Collection<Movie> moviesList = this.movieRepository.findMovieByDirectorsIdDirector(idDirector);
+        if(moviesList.isEmpty()) throw new DirectorNotFoundException("no any director was found");
+        return moviesList;
 
     }
